Revert "Optimize remote state stale file deletion (#13131)"
This reverts commit b9befaa12524744ffe71d3f12699cb27cb8fca5e.
